core_requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile core_requirements.in
#
bcrypt==3.2.0             # via -r core_requirements.in
blinker==1.4              # via flask-mail
cffi==1.14.4              # via -r core_requirements.in, bcrypt
click==7.1.2              # via -r core_requirements.in, flask
flask-mail==0.9.1         # via -r core_requirements.in
flask-sqlalchemy==2.4.4   # via -r core_requirements.in
flask-wtf==0.14.3         # via -r core_requirements.in
flask==1.1.2              # via -r core_requirements.in, flask-mail, flask-sqlalchemy, flask-wtf
itsdangerous==1.1.0       # via -r core_requirements.in, flask, flask-wtf
jinja2==2.11.2            # via -r core_requirements.in, flask
markupsafe==1.1.1         # via -r core_requirements.in, jinja2, wtforms
mysqlclient==2.0.2        # via -r core_requirements.in
pillow==8.0.1             # via -r core_requirements.in
plotly==4.14.1            # via -r core_requirements.in
pycparser==2.20           # via -r core_requirements.in, cffi
python-decouple==3.3      # via -r core_requirements.in
pyyaml==5.3.1             # via -r core_requirements.in
retrying==1.3.3           # via plotly
ruamel.appconfig==0.5.5   # via -r core_requirements.in, yaml-1.3
ruamel.std.argparse==0.8.3  # via -r core_requirements.in, yaml-1.3
six==1.15.0               # via -r core_requirements.in, bcrypt, plotly, retrying
sqlalchemy==1.3.20        # via -r core_requirements.in, flask-sqlalchemy
werkzeug==1.0.1           # via -r core_requirements.in, flask
wtforms==2.3.3            # via flask-wtf
yaml-1.3==0.1.0           # via -r core_requirements.in
